+++ crates/mimic_tests/canister/src/validate.rs
@@ -66,12 +66,12 @@ impl ValidateTester {
     // test_decimal
     fn test_decimal() {
         // OK
-        let d = DecimalMaxDp::from(Decimal::from(1.2));
+        let d = DecimalMaxDp::from(Decimal::new(1, 2));
         let res = validate(&d);
         assert!(res.is_ok(), "{res:?}");
 
         // ERR
-        let d = DecimalMaxDp::from(Decimal::from(1.2453));
+        let d = DecimalMaxDp::from(Decimal::new(1, 15));
         let res = validate(&d);
         assert!(res.is_err(), "{res:?}");
     }